+++ example/test_driver/integration_test.dart
@@ -4,7 +4,11 @@ import 'package:integration_test/integration_test_driver_extended.dart';
 
 Future<void> main() async {
   await integrationDriver(
-    onScreenshot: (String name, List<int> bytes, [Map<String, Object?>? args]) async {
+    onScreenshot: (
+      String name,
+      List<int> bytes, [
+      Map<String, Object?>? args,
+    ]) async {
       final file = File('screenshots/$name.png');
       await file.create(recursive: true);
       await file.writeAsBytes(bytes);